Do not destroy File Vault key on standby
That is tricky thing and it depends on type of machine – should be used
differently on desktop and laptops.

Also we need to make sure additional settings applied:

```
$ sudo pmset -a powernap 0
$ sudo pmset -a standby 0
$ sudo pmset -a standbydelay 0
$ sudo pmset -a autopoweroff 0
```

The worst think on laptops – it's very annoying to enter password twice each time you
close laptop.

Reference: https://github.com/drduh/OS-X-Security-and-Privacy-Guide#full-disk-encryption